packages/backend/pyproject.toml
[build-system]
requires = ["uv-build~=0.9.8"]
build-backend = "uv_build"

[project]
name = "backend"
version = "1.0.0"
requires-python = ">=3.11,<3.14"
classifiers = [
    "Private :: Do Not Upload",
    "Typing :: Typed",
]
dependencies = [
    "asgi-ratelimit==0.10.0",
    "asyncpg==0.30.0",
    "asyncpg-stubs==0.30.2",
    "fastapi[standard]==0.121.1",
    "pwdlib[argon2]==0.3.0",
    "pyjwt[crypto]==2.10.1",
    "uvicorn[standard]==0.38.0",
]
